Prepare for FLTCam migration to Swift
