+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/CameraOrientationTests.swift
@@ -10,7 +10,7 @@ import XCTest
 
 // Import Objectice-C part of the implementation when SwiftPM is used.
 #if canImport(camera_avfoundation_objc)
-  @testable import camera_avfoundation_objc
+  import camera_avfoundation_objc
 #endif
 
 private final class MockUIDevice: UIDevice {
@@ -24,14 +24,14 @@ private final class MockUIDevice: UIDevice {
 final class CameraOrientationTests: XCTestCase {
   private func createCameraPlugin() -> (
     cameraPlugin: CameraPlugin,
-    mockCamera: MockFLTCam,
+    mockCamera: MockCamera,
     mockEventAPI: MockGlobalEventApi,
     mockDevice: MockCaptureDevice,
     mockDeviceDiscoverer: MockCameraDeviceDiscoverer,
     captureSessionQueue: DispatchQueue
   ) {
     let mockDevice = MockCaptureDevice()
-    let mockCamera = MockFLTCam()
+    let mockCamera = MockCamera()
     let mockEventAPI = MockGlobalEventApi()
     let mockDeviceDiscoverer = MockCameraDeviceDiscoverer()
     let captureSessionQueue = DispatchQueue(label: "io.flutter.camera.captureSessionQueue")
