+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/CameraPluginInitializeCameraTests.swift
@@ -8,14 +8,14 @@ import XCTest
 
 // Import Objectice-C part of the implementation when SwiftPM is used.
 #if canImport(camera_avfoundation_objc)
-  @testable import camera_avfoundation_objc
+  import camera_avfoundation_objc
 #endif
 
 final class CameraPluginInitializeCameraTests: XCTestCase {
   private func createCameraPlugin() -> (
-    CameraPlugin, MockFLTCam, MockGlobalEventApi, DispatchQueue
+    CameraPlugin, MockCamera, MockGlobalEventApi, DispatchQueue
   ) {
-    let mockCamera = MockFLTCam()
+    let mockCamera = MockCamera()
     let mockGlobalEventApi = MockGlobalEventApi()
     let captureSessionQueue = DispatchQueue(label: "io.flutter.camera.captureSessionQueue")
