+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/FLTCamExposureTests.swift
@@ -8,7 +8,7 @@ import XCTest
 
 // Import Objectice-C part of the implementation when SwiftPM is used.
 #if canImport(camera_avfoundation_objc)
-  @testable import camera_avfoundation_objc
+  import camera_avfoundation_objc
 #endif
 
 final class FLTCamExposureTests: XCTestCase {
