+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/StreamingTests.swift
@@ -9,7 +9,7 @@ import XCTest
 
 // Import Objectice-C part of the implementation when SwiftPM is used.
 #if canImport(camera_avfoundation_objc)
-  @testable import camera_avfoundation_objc
+  import camera_avfoundation_objc
 #endif
 
 private class MockImageStreamHandler: FLTImageStreamHandler {
@@ -32,19 +32,26 @@ private class MockImageStreamHandler: FLTImageStreamHandler {
 }
 
 final class StreamingTests: XCTestCase {
-  private func createCamera() -> (FLTCam, CMSampleBuffer) {
+  private func createCamera() -> (
+    FLTCam,
+    AVCaptureOutput,
+    CMSampleBuffer,
+    AVCaptureConnection
+  ) {
     let captureSessionQueue = DispatchQueue(label: "testing")
     let configuration = CameraTestUtils.createTestCameraConfiguration()
     configuration.captureSessionQueue = captureSessionQueue
 
     let camera = CameraTestUtils.createTestCamera(configuration)
+    let testAudioOutput = CameraTestUtils.createTestAudioOutput()
     let sampleBuffer = CameraTestUtils.createTestSampleBuffer()
+    let testAudioConnection = CameraTestUtils.createTestConnection(testAudioOutput)
 
-    return (camera, sampleBuffer)
+    return (camera, testAudioOutput, sampleBuffer, testAudioConnection)
   }
 
   func testExceedMaxStreamingPendingFramesCount() {
-    let (camera, sampleBuffer) = createCamera()
+    let (camera, testAudioOutput, sampleBuffer, testAudioConnection) = createCamera()
     let streamingExpectation = expectation(
       description: "Must not call handler over maxStreamingPendingFramesCount")
     let handlerMock = MockImageStreamHandler()
@@ -59,14 +66,14 @@ final class StreamingTests: XCTestCase {
 
     streamingExpectation.expectedFulfillmentCount = 4
     for _ in 0..<10 {
-      camera.captureOutput(nil, didOutputSampleBuffer: sampleBuffer, from: nil)
+      camera.captureOutput(testAudioOutput, didOutput: sampleBuffer, from: testAudioConnection)
     }
 
     waitForExpectations(timeout: 30, handler: nil)
   }
 
   func testReceivedImageStreamData() {
-    let (camera, sampleBuffer) = createCamera()
+    let (camera, testAudioOutput, sampleBuffer, testAudioConnection) = createCamera()
     let streamingExpectation = expectation(
       description: "Must be able to call the handler again when receivedImageStreamData is called")
     let handlerMock = MockImageStreamHandler()
@@ -81,11 +88,11 @@ final class StreamingTests: XCTestCase {
 
     streamingExpectation.expectedFulfillmentCount = 5
     for _ in 0..<10 {
-      camera.captureOutput(nil, didOutputSampleBuffer: sampleBuffer, from: nil)
+      camera.captureOutput(testAudioOutput, didOutput: sampleBuffer, from: testAudioConnection)
     }
 
     camera.receivedImageStreamData()
-    camera.captureOutput(nil, didOutputSampleBuffer: sampleBuffer, from: nil)
+    camera.captureOutput(testAudioOutput, didOutput: sampleBuffer, from: testAudioConnection)
 
     waitForExpectations(timeout: 30, handler: nil)
   }
