+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation_objc/include/camera_avfoundation/FLTCam_Test.h
@@ -8,20 +8,8 @@
 #import "FLTCapturePhotoOutput.h"
 #import "FLTCaptureVideoDataOutput.h"
 #import "FLTDeviceOrientationProviding.h"
+#import "FLTImageStreamHandler.h"
 #import "FLTSavePhotoDelegate.h"
-
-@interface FLTImageStreamHandler : NSObject <FlutterStreamHandler>
-
-/// The queue on which `eventSink` property should be accessed.
-@property(nonatomic, strong) dispatch_queue_t captureSessionQueue;
-
-/// The event sink to stream camera events to Dart.
-///
-/// The property should only be accessed on `captureSessionQueue`.
-/// The block itself should be invoked on the main queue.
-@property FlutterEventSink eventSink;
-
-@end
 
 // APIs exposed for unit testing.
 @interface FLTCam ()
@@ -47,7 +35,8 @@
 /// Exposed for unit tests.
 - (void)captureOutput:(AVCaptureOutput *)output
     didOutputSampleBuffer:(CMSampleBufferRef)sampleBuffer
-           fromConnection:(NSObject<FLTCaptureConnection> *)connection;
+           fromConnection:(AVCaptureConnection *)connection
+    NS_SWIFT_NAME(captureOutput(_:didOutput:from:));
 
 /// Start streaming images.
 - (void)startImageStreamWithMessenger:(NSObject<FlutterBinaryMessenger> *)messenger
